Revert "perf: Replace Double with float"
This reverts commit 3f37765f6351f966dd23d694241a0e75000d90db.

+++ app/src/main/kotlin/com/looker/kenko/data/model/Rating.kt
@@ -18,6 +18,6 @@ import androidx.compose.runtime.Immutable
 
 @Immutable
 @JvmInline
-value class Rating(val value: Float)
+value class Rating(val value: Double)
 
 operator fun Rating.plus(other: Rating) = Rating(value + other.value)
